added eigensystem
